requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
bert-score==0.3.13
blinker==1.8.2
blis==1.1.0
build==1.2.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.7
click==8.1.7
cloudpathlib==0.20.0
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.10
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
durationpy==0.7
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi==0.115.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.56.0
frozenlist==1.4.1
fsspec==2024.9.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
groq==0.22.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
kubernetes==31.0.0
langchain==0.3.0
langchain-chroma==0.1.4
langchain-community==0.3.0
langchain-core==0.3.54
langchain-groq==0.3.2
langchain-huggingface==0.1.2
langchain-ollama==0.2.1
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langcodes==3.5.0
langsmith==0.1.125
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.10.1
mdurl==0.1.2
mmh3==5.0.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.11
mypy-extensions==1.0.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
nvidia-pyindex==1.0.9
oauthlib==3.2.2
ollama==0.4.4
onnxruntime==1.19.2
openai==1.46.1
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
posthog==3.6.6
preshed==3.0.9
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.9.1
pyparsing==3.2.1
pypdf==5.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.1.1
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.5.0
starlette==0.38.5
sympy==1.13.1
tenacity==8.5.0
thinc==8.3.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.1
tomli==2.1.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.66.5
transformers==4.51.3
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
wasabi==1.1.3
watchfiles==0.24.0
weasel==0.4.1
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.11.1
zipp==3.20.2
